+++ full-stack-tests/rpc/src/backend/CommonBackendSetup.ts
@@ -15,9 +15,21 @@ export async function commonSetup(): Promise<void> {
   RpcConfiguration.developmentMode = true;
 
   // Start the backend
-  if (ProcessDetector.isElectronAppBackend)
-    await ElectronHost.startup({ electronHost: { rpcInterfaces } });
-  else
+  if (ProcessDetector.isElectronAppBackend){
+    const baseOidcScopes = [
+      "openid",
+      "email",
+      "profile",
+      "organization",
+      "itwinjs",
+    ];
+    const authConfig = {
+      clientId: "imodeljs-spa-test",
+      redirectUri: "http://localhost:3000/signin-callback",
+      scope: baseOidcScopes.join(" "),
+    };
+    await ElectronHost.startup({ electronHost: { rpcInterfaces, authConfig } });
+  } else
     await IModelHost.startup();
 
   registerBackendCallback(BackendTestCallbacks.registerTestRpcImpl2Class, () => {