+++ test-apps/ui-test-app/src/frontend/appui/imodelopen/IModelOpen.tsx
@@ -12,11 +12,10 @@ import { BlockingPrompt } from "./BlockingPrompt";
 import { IModelList } from "./IModelList";
 import { NavigationItem, NavigationList } from "./Navigation";
 import { ProjectDropdown } from "./ProjectDropdown";
-import { ActivityMessageDetails, ActivityMessageEndReason, AuthorizedFrontendRequestContext, IModelApp } from "@bentley/imodeljs-frontend";
-import { BeDuration } from "@bentley/bentleyjs-core";
+import { ActivityMessageDetails, ActivityMessageEndReason,  IModelApp } from "@bentley/imodeljs-frontend";
+import { AccessToken, BeDuration } from "@bentley/bentleyjs-core";
 import { Button } from "@itwin/itwinui-react";
 import { ITwin, ITwinAccessClient } from "@bentley/context-registry-client";
-import { AccessToken } from "@bentley/itwin-client";
 
 /** Properties for the [[IModelOpen]] component */
 export interface IModelOpenProps {
@@ -73,10 +72,9 @@ export class IModelOpen extends React.Component<IModelOpenProps, IModelOpenState
     if (undefined === token)
       return;
 
-    const ctx = new AuthorizedFrontendRequestContext(token);
-
+    const accessToken = (await IModelApp.authorizationClient?.getAccessToken()) ?? "";
     const client = new ITwinAccessClient();
-    const iTwins = await client.getAll(ctx, { pagination: { skip: 0, top: 10 } });
+    const iTwins = await client.getAll(accessToken, { pagination: { skip: 0, top: 10 } });
     this.setState({
       isLoadingProjects: false,
       isLoadingiModels: true,